requirements.txt
# requirements.txt

asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
crypto==1.4.1
cryptography==44.0.2
daphne==4.1.2
Django==5.1.6
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
msgpack==1.1.0
Naked==0.1.32
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
routing==0.2.0
service-identity==24.2.0
shellescape==3.8.1
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0
zope.interface==7.2
